+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Buffed setup.py file
 - Now freeze team/owner name columns for easier navigating of Excel data
 - Now allow users to overwrite an existing Excel file using a keyword argument
+- Excel sheets now have consistent row colors
 
 ## [1.5.0]